Refactor BLE hooks: unified connection, subscription, and notifications for better modularity
